meta: support nix flake for buliding
As noted in [1] on Linux systems, it is not possible to build phoenixd due to a mismatch in the libc version.

A possible solution is to use Nix and build phoenixd inside the shell with all the dependencies. This way, the host machine can use the binary directly without downgrading libc, which can be dangerous.

[1] https://github.com/ACINQ/phoenixd/issues/1#issuecomment-2018205685
Co-authored-by: Pavol Rusnak <pavol@rusnak.io>
Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>