replace --max-absolute-fee by --max-mining-fee
It makes much more sense to consider only the mining fee, for the absolute
fee check, as it is volatile and amount-independent. The service fee is
in % and predictable. This forced us to use a workaround where we
adjusted `max-absolute-fee` based on `auto-liquidity`.

The default value is the same as before (2% of the default
`auto-liquidity`). Previous option is deprecated and explicitly
rejected.