Update database for new liquidity ads purchase type
Note that we are using a polymorphic serialiser to store the liquidity
purchase data in the database (data type is stored in the blob).

A new received-with db mapping has been added to handle the optional
funding fee type for HTLC payments.

Also fixed the database file name when creating the db drivers, due
to the testnet3/4 changes.