Rework liquidity-ads purchase database storage
Payment details are now stored inside the purchase blob.
The payment details type in the db table is just for
information purpose and is not typed. It is a free text
column.

The column storing the purchase/lease type is now also a
free text column. If it happens to contains legacy lease
data, then we apply a special logic to the blob.

The classes used for storing/versioning liquidity ads data
are moved in a liquidity ads package.

The serialiser for LiquidityAds.FundingFee has been removed
(instead we use the same pattern as the other liquidity data).

The methods from moving to the canonical data to the versioned
db data have been rename: asDb(), asCanonical().