Upgrade to Kotlin 2.1.0 (#145)
Also update a few dependencies.

There is an incompatibility between multiplatform and java plugins, but it's just a warning for now. The suggested fix is to create a module for the java app, but it seems [the issue is addressed in 2.1.20](https://kotlinlang.org/docs/whatsnew-eap.html#kotlin-multiplatform-new-dsl-to-replace-gradle-s-application-plugin).

We are now building sqlite with `cinterop` using a trick found in sqldelight. It gets rids of linking issues on linux x64, because there is no mismatch anymore of glibc versions. We previously had to build the project on an old `Ubuntu 18`.