use ubuntu for docker runtime image
The Alpine base image causes systematic crash for some libsecp256k1
operations. JNI-related issues related to the use of `musl` instead of
`glibc` are well documented.

The issue is fixed by switching alpine for ubuntu, which is
`glibc`-based. Note that it makes the final image significantly larger
(502MB vs 313MB).

The goal of using the JVM on docker was for compatibility with ARM. Once
we support native ARM on Linux (#157), we could potentially move back to
Alpine with native phoenixd, an even smaller footprint than before.

Fixes #159.