make fee credit default value dynamic
The fee credit must be significantly greater than the typical fee
for the configured `auto-liquidity` amount, otherwise we will reject
payments that could go through.

The reason is that liquidity fees depend on mining fees which are
volatile, and may increase between the time when phoenix accepts otf
liquidity, and the time when the LSP allocates the liquidity. We want
phoenix to be able to afford the increase of cost, so we require the
incoming payment + fee credit to be tx the estimated fee.

But this creates another issue: if the max allowed fee credit is less
than the minimal liquidity fee (which was the case with the fixed default
of 100k, if the user sets `auto-liquidity=10m`), then we will end up
rejecting an incoming payment that is comfortably above, but less than
two times, the total liquidity fee.