do not filter out events with negative amounts
Depending on liquidity settings, it may happen that a small incoming
payment triggers a large purchase of inbound liquidity, paying from the
current channel balance. In this scenario, the received amount for that
particular will be negative, because the cost of the purchase is greater
than the value of that payment. We should still emit the associated
webhook/websocket event.